+++ CHANGELOG.md
@@ -3,10 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## [0.0.2] - 2017-10-23
-### Wysiwyg images in Windows
+## [0.0.2] - 2017-10-24
+### Multi Stores
 
-- store | detect store/website by url without ```index.php``` or ```htaccess/```
+- store | detect store/website by url without ```index.php``` or ```.htaccess```
 
 ## [0.0.1] - 2017-09-19
 ### Wysiwyg images in Windows